// generated/src/app/api-models/prem-bill-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PremBillSetup {
  seqGpbilId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  requestType: string;
  action: string;
  billingBatch: string;
  billingMonth: Date;
  billingType: string;
  reportFormat: string;
  postMonth: Date;
  reprintRequest: string;
  template: string;
  daemonRequest: string;
  status: string;
  comments: string;
  inProcess: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  subscFlag: string;
}